+++ docs/source/conf.py
@@ -359,7 +359,7 @@ autodoc_mock_imports = [
 # Example configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3/", None),
-    "cg": ("https://cryptography.io/en/%s/" % CONFIG["cryptography-major"][-1], None),
+    "cg": ("https://cryptography.io/en/%s/" % CONFIG["cryptography"][-1], None),
     "django": (
         "https://docs.djangoproject.com/en/%s/" % CONFIG["django-major"][-1],
         "https://docs.djangoproject.com/en/%s/_objects/" % CONFIG["django-major"][-1],